+++ src/FakeXrmEasy.Plugins/Pipeline/Scope/MaxDepthExceededException.cs
@@ -1,0 +1,7 @@
+namespace FakeXrmEasy.Pipeline.Scope
+{
+    public class MaxDepthExceededException
+    {
+        
+    }
+}