+++ src/FakeXrmEasy.Plugins/Pipeline/Scope/PipelineOrganizationRequest.cs
@@ -5,7 +5,7 @@ namespace FakeXrmEasy.Pipeline.Scope
     /// <summary>
     /// An internal organization request that decorates additional properties for pipeline processing
     /// </summary>
-    internal class PipelineOrganizationRequest
+    internal class PipelineOrganizationRequest: OrganizationRequest
     {
         /// <summary>
         /// A reference to the the original request that is being executed