+++ src/FakeXrmEasy.Plugins/Pipeline/Scope/PipelineOrganizationServiceFactory.cs
@@ -1,6 +1,7 @@
 using System;
 using FakeItEasy;
 using Microsoft.Xrm.Sdk;
+using Microsoft.Xrm.Sdk.Messages;
 using Microsoft.Xrm.Sdk.Query;
 
 namespace FakeXrmEasy.Pipeline.Scope
@@ -30,55 +31,142 @@ namespace FakeXrmEasy.Pipeline.Scope
             AddFakeRetrieve(pipelineService, service);
             AddFakeAssociate(pipelineService, service);
             AddFakeDisassociate(pipelineService, service);
+            AddFakeExecute(pipelineService, service);
         }
         
         private static void AddFakeCreate(IPipelineOrganizationService pipelineService, IOrganizationService service) 
         {
             A.CallTo(() => pipelineService.Create(A<Entity>._))
                 .ReturnsLazily((Entity e) =>
-                    service.Create(e));
+                {
+                    var response = service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = new CreateRequest() { Target = e },
+                        CurrentScope = pipelineService.Scope
+                    });
+                    return (response as CreateResponse).id;
+                });
         }
         
-        private static void AddFakeUpdate(IPipelineOrganizationService pipelineService, IOrganizationService service) 
+        private static void AddFakeUpdate(IPipelineOrganizationService pipelineService, IOrganizationService service)
         {
             A.CallTo(() => pipelineService.Update(A<Entity>._))
                 .Invokes((Entity e) =>
-                    service.Update(e));
+                {
+                    service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = new UpdateRequest() { Target = e },
+                        CurrentScope = pipelineService.Scope
+                    });
+                });
         }
         
         private static void AddFakeRetrieve(IPipelineOrganizationService pipelineService, IOrganizationService service)
         {
             A.CallTo(() => pipelineService.Retrieve(A<string>._, A<Guid>._, A<ColumnSet>._))
                 .ReturnsLazily((string entityName, Guid id, ColumnSet columnSet) 
-                    => service.Retrieve(entityName, id, columnSet));
+                    =>
+                {
+                    var response = service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = new RetrieveRequest()
+                        {
+                            Target = new EntityReference(entityName, id),
+                            ColumnSet = columnSet
+                        },
+                        CurrentScope = pipelineService.Scope
+                    });
+                    
+                    return (response as RetrieveResponse).Entity;
+                });
         }
         
         private static void AddFakeRetrieveMultiple(IPipelineOrganizationService pipelineService, IOrganizationService service)
         {
             A.CallTo(() => pipelineService.RetrieveMultiple(A<QueryBase>._))
                 .ReturnsLazily((QueryBase req) =>
-                    service.RetrieveMultiple(req));
+                {
+                    var response = service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = new RetrieveMultipleRequest()
+                        {
+                            Query = req
+                        },
+                        CurrentScope = pipelineService.Scope
+                    });
+
+                    return (response as RetrieveMultipleResponse).EntityCollection;
+                });
         }
         
         private static void AddFakeDelete(IPipelineOrganizationService pipelineService, IOrganizationService service)
         {
             A.CallTo(() => pipelineService.Delete(A<string>._, A<Guid>._))
                 .Invokes((string entityName, Guid id) =>
-                    service.Delete(entityName, id));
+                {
+                    service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = new DeleteRequest()
+                        {
+                            Target = new EntityReference(entityName, id)
+                        },
+                        CurrentScope = pipelineService.Scope
+                    });
+                });
         }
         
         private static void AddFakeAssociate(IPipelineOrganizationService pipelineService, IOrganizationService service)
         {
-            A.CallTo(() => pipelineService.Associate(A<string>._, A<Guid>._, A<Relationship>._, A<EntityReferenceCollection>._))
-                .Invokes((string entityName, Guid entityId, Relationship relationship, EntityReferenceCollection entityCollection) =>
-                    service.Associate(entityName, entityId, relationship, entityCollection));
+            A.CallTo(() =>
+                    pipelineService.Associate(A<string>._, A<Guid>._, A<Relationship>._,
+                        A<EntityReferenceCollection>._))
+                .Invokes((string entityName, Guid entityId, Relationship relationship,
+                    EntityReferenceCollection entityCollection) =>
+                {
+                    service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = new AssociateRequest()
+                        {
+                            Target = new EntityReference(entityName, entityId),
+                            Relationship = relationship,
+                            RelatedEntities = entityCollection
+                        },
+                        CurrentScope = pipelineService.Scope
+                    });
+                });
         }
         
         private static void AddFakeDisassociate(IPipelineOrganizationService pipelineService, IOrganizationService service)
         {
-            A.CallTo(() => pipelineService.Disassociate(A<string>._, A<Guid>._, A<Relationship>._, A<EntityReferenceCollection>._))
-                .Invokes((string entityName, Guid entityId, Relationship relationship, EntityReferenceCollection entityCollection) =>
-                    service.Disassociate(entityName, entityId, relationship, entityCollection));
+            A.CallTo(() =>
+                    pipelineService.Disassociate(A<string>._, A<Guid>._, A<Relationship>._,
+                        A<EntityReferenceCollection>._))
+                .Invokes((string entityName, Guid entityId, Relationship relationship,
+                    EntityReferenceCollection entityCollection) =>
+                {
+                    service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = new DisassociateRequest()
+                        {
+                            Target = new EntityReference(entityName, entityId),
+                            Relationship = relationship,
+                            RelatedEntities = entityCollection
+                        },
+                        CurrentScope = pipelineService.Scope
+                    });
+                });
+        }
+        
+        private static void AddFakeExecute(IPipelineOrganizationService pipelineService, IOrganizationService service)
+        {
+            A.CallTo(() => pipelineService.Execute(A<OrganizationRequest>._))
+                .ReturnsLazily((OrganizationRequest request) => 
+                    service.Execute(new PipelineOrganizationRequest()
+                    {
+                        OriginalRequest = request,
+                        CurrentScope = pipelineService.Scope
+                    })
+                );
         }
     }
 }